rename: 상품 삭제 컨테이너 파일명 변경
DeleteProductButton, DeleteProductIconButton으로 변경한다.
