+++ src/containers/DeleteProductIconButton/index.jsx
