+++ src/stories/Button.stories.jsx
@@ -5,10 +5,10 @@ import HomeButton from 'containers/HomeButton';
 import OrderListButton from 'containers/OrderListButton';
 import MoveCartPageButton from 'containers/MoveCartPageButton';
 import DeleteProductButton from 'containers/DeleteProductButton';
+import DeleteProductIconButton from 'containers/DeleteProductIconButton';
 import OrderProductsButton from 'containers/OrderProductsButton';
 import ProductCountUpButton from 'containers/ProductCountUpButton';
 import ProductCountDownButton from 'containers/ProductCountDownButton';
-import ProductDeleteButton from 'containers/ProductDeleteButton';
 
 export default {
   title: 'containers/Button',
@@ -23,17 +23,17 @@ const DeleteProductButtonTemplate = () => <DeleteProductButton />;
 const OrderProductsButtonTemplate = () => <OrderProductsButton />;
 const ProductCountUpButtonTemplate = () => <ProductCountUpButton />;
 const ProductCountDownButtonTemplate = () => <ProductCountDownButton />;
-const ProductDeleteButtonTemplate = () => <ProductDeleteButton />;
+const DeleteProductIconButtonTemplate = () => <DeleteProductIconButton />;
 
 export const Primary = Template.bind({});
 export const HomeButtonPrimary = HomeButtonTemplate.bind({});
 export const OrderListButtonPrimary = OrderListButtonTemplate.bind({});
 export const MoveCartPageButtonPrimary = MoveCartPageButtonTemplate.bind({});
 export const DeleteProductButtonPrimary = DeleteProductButtonTemplate.bind({});
+export const DeleteProductIconButtonPrimary = DeleteProductIconButtonTemplate.bind({});
 export const OrderProductsButtonPrimary = OrderProductsButtonTemplate.bind({});
 export const ProductCountUpButtonPrimary = ProductCountUpButtonTemplate.bind({});
 export const ProductCountDownButtonPrimary = ProductCountDownButtonTemplate.bind({});
-export const ProductDeleteButtonPrimary = ProductDeleteButtonTemplate.bind({});
 
 Primary.args = {
   width: '150px',
